extract control creation from main bar-chart method
